requirements.txt
asgiref==3.11.1
Django==6.0.2
pillow==12.1.1
razorpay==1.4.1
sqlparse==0.5.5
tzdata==2025.3
twilio==9.10.3
python-dotenv==1.2.2
